@@ -2,6 +2,7 @@ import { request } from '../helpers/request'
 
 import { GET_CURRENT_USER, GET_RESERVABLE_GARAGES } from '../queries/mobile.header.queries'
 import { REVOKE_TOKEN } from '../queries/login.queries'
+import { t } from '../modules/localization/localization'
 
 
 export const MOBILE_MENU_SET_GARAGES = 'MOBILE_MENU_SET_GARAGES'
@@ -10,6 +11,7 @@ export const MOBILE_MENU_SET_CURRENT_USER = 'MOBILE_MENU_SET_CURRENT_USER'
 export const MOBILE_MENU_SET_SHOW_MENU = 'MOBILE_MENU_SET_SHOW_MENU'
 export const MOBILE_MENU_SET_ERROR = 'MOBILE_MENU_SET_ERROR'
 export const MOBILE_MENU_SET_CUSTOM_MODAL = 'MOBILE_MENU_SET_CUSTOM_MODAL'
+export const SET_MOBILE_LANGUAGE = 'SET_MOBILE_LANGUAGE'
 
 
 export function resetStore() {
@@ -17,40 +19,52 @@ export function resetStore() {
 }
 
 export function setGarages(garages) {
-  return { type: MOBILE_MENU_SET_GARAGES
-         , value: garages
-         }
+  return {
+    type:  MOBILE_MENU_SET_GARAGES,
+    value: garages
+  }
 }
 
 export function setGarage(garage) {
-  return { type: MOBILE_MENU_SET_GARAGE
-         , value: garage
-         }
+  return {
+    type:  MOBILE_MENU_SET_GARAGE,
+    value: garage
+  }
 }
 
-
 export function setCurrentUser(currentUser) {
-  return { type: MOBILE_MENU_SET_CURRENT_USER
-         , value: currentUser
-         }
+  return {
+    type:  MOBILE_MENU_SET_CURRENT_USER,
+    value: currentUser
+  }
 }
 
 export function setShowMenu(bool) {
-  return { type: MOBILE_MENU_SET_SHOW_MENU
-         , value: bool
-         }
+  return {
+    type:  MOBILE_MENU_SET_SHOW_MENU,
+    value: bool
+  }
 }
 
 export function setError(text) {
-  return { type: MOBILE_MENU_SET_ERROR
-         , value: text
-         }
+  return {
+    type:  MOBILE_MENU_SET_ERROR,
+    value: text
+  }
 }
 
 export function setCustomModal(content) {
-  return { type: MOBILE_MENU_SET_CUSTOM_MODAL
-         , value: content
-         }
+  return {
+    type:  MOBILE_MENU_SET_CUSTOM_MODAL,
+    value: content
+  }
+}
+
+export function setLanguage(langugage) {
+  return {
+    type:  SET_MOBILE_LANGUAGE,
+    value: langugage
+  }
 }
 
 
@@ -58,11 +72,12 @@ export function initGarages() {
   return dispatch => {
     const onGarageSuccess = response => {
       const garages = response.data.reservable_garages
-      garages.unshift({ id: undefined, name: 'All garages' })
+      garages.unshift({ id: undefined, name: t([ 'mobileApp', 'page', 'allGarages' ]) })
       dispatch(setGarages(garages))
     }
 
     const onSuccess = response => {
+      dispatch(setLanguage(response.data.current_user.language))
       dispatch(setCurrentUser(response.data.current_user))
       request(onGarageSuccess, GET_RESERVABLE_GARAGES, { user_id: response.data.current_user.id })
     }
