@@ -1,4 +1,6 @@
-import React from 'react'
+import React       from 'react'
+import { connect } from 'react-redux'
+import { confirm } from '../../actions/pageBase.actions'
 
 // Super of all buttons
 // content = ...
@@ -9,12 +11,13 @@ import React from 'react'
 // question = text of confirm window when type == 'remove'
 
 
-export default function Button({ content, onClick, onDisabledClick, type, state, style, question }) {
+function Button({ content, onClick, onDisabledClick, type, state, style, question, confirm }) {
   const handleClick = e => {
     e.stopPropagation()
     if (typeof onClick === 'function') { // if no fuction, do nothing
       if (type === 'remove') {
-        confirm(question || 'Are you sure?') && onClick()
+        // confirm(question || 'Are you sure?') && onClick()
+        confirm(question || 'Are you sure?', onClick)
       } else {
         onClick()
       }
@@ -25,3 +28,8 @@ export default function Button({ content, onClick, onDisabledClick, type, state,
     <button className={style} onClick={state === 'disabled' ? onDisabledClick : handleClick} type="button">{content}</button>
   )
 }
+
+export default connect(
+  () => ({}),
+  dispatch => ({ confirm: (question, onClick) => dispatch(confirm(question, onClick)) })
+)(Button)
