@@ -8,10 +8,10 @@ import Button from './Button.js'
 // state = 'selected', 'disabled'
 
 
-export default function CallToActionButton({ label, onClick, state}) {
-  const style = [
-    styles.button,
-    styles[state]
+export default function CallToActionButton({ label, onClick, state, type }) {
+  const style = [ styles.button,
+    styles[state],
+    styles[type]
   ].join(' ')
 
   const content = <span className={styles.label}>{label}</span>
