@@ -80,7 +80,7 @@ class PageBase extends Component {
                           </div>
 
     const notificationsModal = <div style={{"textAlign": "center"}}>
-                                {t(['pageBase', 'unredNotifications'], {count: notifications.count})}: <br/>
+                                {t(['pageBase', 'unredNotifications'], {count: notifications.count})}. <br/>
                                 <RoundButton content={<i className="fa fa-check" aria-hidden="true"></i>} onClick={notificationsModalClick} type='confirm'  />
                               </div>
 
