@@ -1,14 +1,20 @@
-import translate       from 'counterpart'
+import translate from 'counterpart'
 import { hashHistory } from 'react-router'
+import { composeParameters, parseParameters } from './parseUrlParameters'
 
-export function to (dest) {
+
+export function path(dest) {
+  return `/${translate.getLocale()}${dest}`
+}
+
+export function to(dest) {
   hashHistory.push(path(dest))
 }
 
-export function path (dest) {
-  return `${translate.getLocale()}${dest}`
+export function back() {
+  hashHistory.goBack()
 }
 
-export function back (){
-  hashHistory.goBack()
+export function changeLanguage(language) {
+  hashHistory.push(`${language}${window.location.hash.substr(4).split('?')[0]}?${composeParameters(parseParameters(window.location.hash))}`)
 }
