@@ -1,14 +1,10 @@
-'use strict';
+module.exports = function pluralizeDe(entry, count) {
+  // 'one' - 1 day
+  if (count === 1) { return entry.one }
 
-// Specify when to use keys 'one', 'few', 'many', 'other'
-// 'one' - 1 day
-// 'few' - nothing for english
-// 'many' - nothing for english
-// 'other' - 5 days
-module.exports = function(entry, count) {
-  if (count == 1) {
-    return entry['one']
-  }
+  // 'few' - nothing for de
+  // 'many' - nothing for de
 
-  return entry['other']
-};
+  // 'other' - 5 days
+  return entry.other
+}
