@@ -1,14 +1,15 @@
-'use strict';
+module.exports = function pluralizePl(entry, count) {
+  const mod10 = count % 10
+  const mod100 = count % 100
 
-// Specify when to use keys 'one', 'few', 'many', 'other'
-// 'one' - 1 day
-// 'few' - nothing for english
-// 'many' - nothing for english
-// 'other' - 5 days
-module.exports = function(entry, count) {
-  if (count == 1) {
-    return entry['one']
-  }
+  // count is 1
+  if (count === 1) { return entry.one }
 
-  return entry['other']
-};
+  // count ends with 2, 3 or 4 but does not end with 11, 12, 13
+  if ([ 2, 3, 4 ].includes(mod10) && ![ 12, 13, 14 ].includes(mod100)) { return entry.few }
+
+  // count ends with 0, 1, 5 - 9 or ends with 11, 12, 13
+  if ([ 0, 1, 5, 6, 7, 8, 9 ].includes(mod10) || [ 12, 13, 14 ].includes(mod100)) { return entry.many }
+
+  return entry.other
+}
