@@ -1,66 +1,88 @@
+import * as localization from '../modules/localization/localization'
+
 import {
   MOBILE_MENU_SET_GARAGES,
   MOBILE_MENU_SET_GARAGE,
   MOBILE_MENU_SET_CURRENT_USER,
   MOBILE_MENU_SET_SHOW_MENU,
   MOBILE_MENU_SET_ERROR,
-  MOBILE_MENU_SET_CUSTOM_MODAL
+  MOBILE_MENU_SET_CUSTOM_MODAL,
+  SET_MOBILE_LANGUAGE
 } from '../actions/mobile.header.actions'
 
 import { SET_NOTIFICATIONS_COUNT } from '../actions/notifications.actions'
 
-const defaultState = { garages: []
-                     , garage_id: undefined
-                     , current_user: undefined
-                     , headerHeight: 101 // px
-                     , online: navigator.connection ? navigator.connection.type !== 'none' : true
-                     , showMenu: false
-                     , error: undefined
-                     , custom_modal: undefined
-                     , notificationsCount: 0
-                     }
+const defaultState = {
+  garages:            [],
+  garage_id:          undefined,
+  current_user:       undefined,
+  online:             navigator.connection ? navigator.connection.type !== 'none' : true,
+  showMenu:           false,
+  error:              undefined,
+  custom_modal:       undefined,
+  notificationsCount: 0,
+  language:           'en'
+}
+//  , headerHeight: 101 // px
 
 
 export default function mobileHeader(state = defaultState, action) {
   switch (action.type) {
-
     case MOBILE_MENU_SET_GARAGES:
-      return { ...state
-             , garages: action.value
-             }
+      return {
+        ...state,
+        garages: action.value
+      }
 
     case MOBILE_MENU_SET_GARAGE:
-      return { ...state
-             , garage_id: action.value
-             }
+      return {
+        ...state,
+        garage_id: action.value
+      }
 
     case MOBILE_MENU_SET_CURRENT_USER:
-      return { ...state
-             , current_user: action.value
-             }
+      return {
+        ...state,
+        current_user: action.value
+      }
 
     case MOBILE_MENU_SET_SHOW_MENU:
-      return { ...state
-             , showMenu: action.value
-             }
+      return {
+        ...state,
+        showMenu: action.value
+      }
+
     case MOBILE_MENU_SET_ERROR:
-      return { ...state
-             , error: action.value
-             }
+      return {
+        ...state,
+        error: action.value
+      }
+
     case MOBILE_MENU_SET_CUSTOM_MODAL:
-      return { ...state
-             , custom_modal: action.value
-             }
+      return {
+        ...state,
+        custom_modal: action.value
+      }
 
     case SET_NOTIFICATIONS_COUNT:
-      return { ...state
-             , notificationsCount: action.value
-             }
+      return {
+        ...state,
+        notificationsCount: action.value
+      }
+
+
+    case SET_MOBILE_LANGUAGE:
+      localization.changeLanguage(action.value)
+      return {
+        ...state,
+        language: action.value
+      }
 
     case 'MOBIE_MENU_SET_DEVICE_ONLINE':
-      return { ...state
-             , online: action.value
-             }
+      return {
+        ...state,
+        online: action.value
+      }
 
     default:
       return state
