@@ -1,9 +1,20 @@
+import moment from 'moment'
+import { MOMENT_DATETIME_FORMAT, MOMENT_DATE_FORMAT } from '../helpers/time'
+
 import {
   NEW_RESERVATION_SET_USER,
   NEW_RESERVATION_SET_AVAILABLE_USERS,
   NEW_RESERVATION_SET_RESERVATION,
 
+  NEW_RESERVATION_SET_HOST_NAME,
+  NEW_RESERVATION_SET_HOST_PHONE,
+  NEW_RESERVATION_SET_HOST_EMAIL,
+
   NEW_RESERVATION_SET_CLIENT_ID,
+  NEW_RESERVATION_SET_RECURRING_RULE,
+  NEW_RESERVATION_SHOW_RECURRING,
+  NEW_RESERVATION_SET_USE_RECURRING,
+  NEW_RESERVATION_SET_RECURRING_RESERVATION_ID,
 
   NEW_RESERVATION_CAR_ID,
   NEW_RESERVATION_CAR_LICENCE_PLATE,
@@ -23,111 +34,167 @@ import {
   NEW_RESERVATION_CLEAR_FORM
 }  from '../actions/newReservation.actions'
 
-const defaultState =  { user:             undefined // id of selected user reservation is for
-                      , availableUsers:   [] // array of other available users
-                      , reservation:      undefined // object with reservation to be edited
+const defaultState = {
+  user:           undefined, // id of selected user reservation is for
+  availableUsers: [], // array of other available users
+  reservation:    undefined, // object with reservation to be edited
 
-                      , client_id:        undefined // currently selected client
+  name:  { value: '', valid: false },
+  phone: { value: '', valid: false },
+  email: { value: '', valid: false },
 
-                      , car_id:           undefined // selected car id
-                      , carLicencePlate:  '' // in case there are no available cars
+  client_id:                undefined, // currently selected client
+  recurringRule:            undefined, // rule in string
+  showRecurring:            false, // show recurring modal
+  useRecurring:             false, // use recurring
+  recurring_reservation_id: undefined,
 
-                      , garage:           undefined // loaded garage details // GARAGE ID IS IN HERE
+  car_id:          undefined, // selected car id
+  carLicencePlate: '', // in case there are no available cars
 
-                      , from:             ''
-                      , to:               ''
-                      , place_id:         undefined // if of selected place
-                      , price:            undefined
+  garage: undefined, // loaded garage details // GARAGE ID IS IN HERE
 
-                      , durationDate:     false // set duration or end of parking?
-                      , loading:          false
-                      , hightlight:       false
-                      , error:            undefined
-                      }
+  from:     '',
+  to:       '',
+  place_id: undefined, // if of selected place
+  price:    undefined,
+
+  durationDate: false, // set duration or end of parking?
+  loading:      false,
+  hightlight:   false,
+  error:        undefined
+}
 
 
-export default function newReservation (state = defaultState, action) {
+export default function newReservation(state = defaultState, action) {
   switch (action.type) {
 
     case NEW_RESERVATION_SET_USER:
-      return { ...state
-             , user: action.value
-             }
+      return { ...state,
+        user: action.value
+      }
 
     case NEW_RESERVATION_SET_AVAILABLE_USERS:
-      return { ...state
-             , availableUsers: action.value
-             }
+      return { ...state,
+        availableUsers: action.value
+      }
 
     case NEW_RESERVATION_SET_RESERVATION:
-      return { ...state
-             , reservation: action.value
-             }
+      return { ...state,
+        reservation: action.value
+      }
 
+
+    case NEW_RESERVATION_SET_HOST_NAME:
+      return { ...state,
+        name: action.value
+      }
+
+    case NEW_RESERVATION_SET_HOST_PHONE:
+      return { ...state,
+        phone: action.value
+      }
+
+    case NEW_RESERVATION_SET_HOST_EMAIL:
+      return { ...state,
+        email: action.value
+      }
 
     case NEW_RESERVATION_SET_CLIENT_ID:
-      return { ...state
-             , client_id: action.value
-             }
+      return { ...state,
+        client_id:     action.value,
+        recurringRule: action.value ? state.recurringRule : undefined
+      }
 
+    case NEW_RESERVATION_SET_RECURRING_RULE: {
+      const overMonth = moment(state.to, MOMENT_DATETIME_FORMAT).diff(moment(state.from, MOMENT_DATETIME_FORMAT), 'months') >= 1
+      return { ...state,
+        recurringRule: action.value,
+        showRecurring: false,
+        useRecurring:  !overMonth && !!action.value
+      }
+    }
+
+    case NEW_RESERVATION_SHOW_RECURRING:
+      return { ...state,
+        showRecurring: action.value
+      }
+
+    case NEW_RESERVATION_SET_USE_RECURRING: {
+      const overMonth = moment(state.to, MOMENT_DATETIME_FORMAT).diff(moment(state.from, MOMENT_DATETIME_FORMAT), 'months') >= 1
+      return { ...state,
+        useRecurring:  !overMonth && action.value,
+        showRecurring: !overMonth && action.value && !state.rule
+      }
+    }
+
+    case NEW_RESERVATION_SET_RECURRING_RESERVATION_ID: {
+      return { ...state,
+        recurring_reservation_id: action.value
+      }
+    }
 
     case NEW_RESERVATION_CAR_ID:
-      return { ...state
-             , car_id: action.value
-             }
+      return { ...state,
+        car_id: action.value
+      }
 
     case NEW_RESERVATION_CAR_LICENCE_PLATE:
-      return { ...state
-             , carLicencePlate: action.value
-             }
+      return { ...state,
+        carLicencePlate: action.value
+      }
 
 
     case NEW_RESERVATION_SET_GARAGE:
-      return { ...state
-             , garage: action.value
-             }
+      return { ...state,
+        garage: action.value
+      }
 
 
     case NEW_RESERVATION_SET_FROM:
-      return { ...state
-             , from: action.value
-             }
+      return { ...state,
+        from:          action.value,
+        to:            action.to || state.to,
+        recurringRule: { ...state.recurringRule,
+          starts: moment(action.value, MOMENT_DATETIME_FORMAT).format(MOMENT_DATE_FORMAT)
+        }
+      }
 
     case NEW_RESERVATION_SET_TO:
-      return { ...state
-             , to: action.value
-             }
+      return { ...state,
+        to: action.value
+      }
 
     case NEW_RESERVATION_SET_PLACE_ID:
-      return { ...state
-             , place_id: action.value
-             }
+      return { ...state,
+        place_id: action.value
+      }
 
     case NEW_RESERVATION_SET_PRICE:
-      return { ...state
-             , price: action.value
-             }
+      return { ...state,
+        price: action.value
+      }
 
 
     case NEW_RESERVATION_SET_DURATION_DATE:
-      return { ...state
-             , durationDate: action.value
-             }
+      return { ...state,
+        durationDate: action.value
+      }
 
     case NEW_RESERVATION_SET_LOADING:
-      return { ...state
-             , loading: action.value
-             }
+      return { ...state,
+        loading: action.value
+      }
 
     case NEW_RESERVATION_SET_HIGHLIGHT:
-      return { ...state
-             , highlight: action.value
-             }
+      return { ...state,
+        highlight: action.value
+      }
 
     case NEW_RESERVATION_SET_ERROR:
-      return { ...state
-             , error: action.value
-             }
+      return { ...state,
+        error: action.value
+      }
 
 
     case NEW_RESERVATION_CLEAR_FORM:
