@@ -4,6 +4,7 @@ import { bindActionCreators }          from 'redux'
 
 import * as nav                        from '../../_shared/helpers/navigation'
 import { t }                           from '../../_shared/modules/localization/localization'
+import { AVAILABLE_LANGUAGES }         from '../../routes'
 
 import PageBase     from '../../_shared/containers/pageBase/PageBase'
 import Dropdown     from '../../_shared/components/dropdown/Dropdown'
@@ -40,7 +41,7 @@ class inviteUserPage extends Component {
 
     const checkSubmitable = () => {
       if (!state.email.valid) return false
-      if (!/\+?\(?\d{2,4}\)?[\d\s-]{3,}/.test(state.phone) && state.phone!=="" ) return false
+      if (!/\+[\d]{2,4}[\d]{3,}/.test(state.phone) && state.phone!=="" ) return false
       if (!/^(?!\s*$).+/.test(state.full_name) && state.phone!=="" ) return false
       return true
     }
@@ -80,6 +81,10 @@ class inviteUserPage extends Component {
                              { state.currentEmail }
                            </div>
 
+   const renderLanguage = lang => <span className={state.language === lang ? styles.boldText : styles.inactiveText} onClick={() => { actions.setLanguage(lang) }}>{t(['languages', lang])}</span>
+   const separator = <span>|</span>
+   const addSeparator = (acc, lang, index, array) => array.length-1 !== index ? [ ...acc, lang, separator ] : [ ...acc, lang ]
+
     return (
       <PageBase>
         <Modal content={errorContent} show={state.error!=undefined} />
@@ -105,7 +110,11 @@ class inviteUserPage extends Component {
             <h3>{t(['inviteUser', 'optionalSettings'])}</h3>
             <p>{t(['inviteUser', 'optionalSettingsText'])}</p>
             <PatternInput onEnter={submitForm} onChange={nameChanged} label={t(['inviteUser', 'nameLabel'])} error={t(['signup_page', 'nameInvalid'])} pattern="^(?!\s*$).+" value={state.full_name} />
-            <PatternInput onEnter={submitForm} onChange={phoneChanged} label={t(['inviteUser', 'phoneLabel'])} error={t(['signup_page', 'phoneInvalid'])} pattern="\+?[\d]{3,}" value={state.phone} />
+            <PatternInput onEnter={submitForm} onChange={phoneChanged} label={t(['inviteUser', 'phoneLabel'])} error={t(['signup_page', 'phoneInvalid'])} pattern="\+[\d]{2,4}[\d]{3,}" value={state.phone} />
+            <div>
+              <h3>{t(['languages', 'language'])}</h3>
+              {AVAILABLE_LANGUAGES.map(renderLanguage).reduce(addSeparator, [])}
+            </div>
             {state.client_id && <div>
               <h3>{t(['inviteUser', 'clientRights'])}</h3>
               <p>{t(['inviteUser', 'clientRightsDesc'])}</p>
