@@ -8,6 +8,7 @@ import Modal        from '../_shared/components/modal/Modal'
 import RoundButton  from '../_shared/components/buttons/RoundButton'
 import PatternInput from  '../_shared/components/input/PatternInput'
 import Form         from '../_shared/components/form/Form'
+import Localization from '../_shared/components/localization/Localization'
 
 import * as nav           from '../_shared/helpers/navigation'
 import { t }              from '../_shared/modules/localization/localization'
@@ -19,7 +20,7 @@ const MINIMUM_PASSWORD_LENGTH = 4
 
 
 const NAME_REGEX  = `^(?!\\s*$).+`
-const PHONE_REGEX = `\\+?[\\d]{3,}`
+const PHONE_REGEX = `\\+[\\d]{2,4}[\\d]{3,}`
 const EMAIL_REGEX = `[a-zA-Z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,3}$`
 
 
@@ -54,6 +55,7 @@ class SignUpPage extends Component {
 
     const content = <div>
                       <Logo style='round'/>
+                      <Localization />
 
                       <Modal content={state.fetching ? loadingContent : errorContent} show={state.fetching||state.error!=undefined} />
                       <div className={styles.signUpPage}>
@@ -65,6 +67,7 @@ class SignUpPage extends Component {
                           <PatternInput onEnter={onSubmit} onChange={actions.setConfirmation} label={t(['signup_page', 'confirmation'])}  error={t(['signup_page', 'noMatching'])}                                type='password' pattern={state.password.value}                        value={state.confirmation.value} />
                         </Form>
                       </div>
+
                     </div>
 
     return (
